refactor(SpokeUtils): Relocate validateFillForDeposit
This function operates solely on a high-level objects (i.e. RelayData +
messageHash) and there is no chain-specific implementation. Relocate it
to DepositUtils because SpokeUtils will become chain-specific.
